refactor(pci-rancher): rename page file according naming convention ref: TAPC-987
