--- packages/manager/apps/pci-rancher/src/pages/_app.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-
-const App: React.FC<React.PropsWithChildren> = ({ children }) => {
-  return <>{children}</>;
-};
-export default App;
