+++ packages/manager/apps/pci-rancher/src/pages/home/Home.page.tsx
@@ -6,7 +6,7 @@ import Breadcrumb, {
 } from '@/components/Breadcrumb/Breadcrumb.component';
 import Loading from '@/components/Loading/Loading.component';
 import { useRanchers } from '@/data/hooks/useRancher/useRancher';
-import Listing from './listing';
+import Listing from '../listing/Listing.page';
 
 export function breadcrumb({ params }: BreadcrumbHandleParams) {
   return params.serviceName;
